requirements.txt
pyserial==3.4
Flask==1.0.2
